refactored logic from route into controller
