+++ server/controllers/job.controller.ts
@@ -1,25 +1,41 @@
 // External Dependencies
-import * as mongoDB from "mongodb";
-import * as dotenv from "dotenv";
+import { Request, Response } from "express";
+import { ObjectId } from "mongodb";
+import { collections } from "../src/database/conn";
+import Job from "../models/job";
 
-// Global Variables
-export const collections: { jobs?: mongoDB.Collection } = {}
+const getJob = async (req: Request, res: Response) => {
+    const id = req?.params?.id;
 
-// Initialize Connection
-export async function connectToDatabase () {
-    dotenv.config();
-    
-     // TODO: fix DB_CONN_STRING
-    const client: mongoDB.MongoClient = new mongoDB.MongoClient(process.env.DB_CONN_STRING);
-            
-    await client.connect();
-        
-    const db: mongoDB.Db = client.db(process.env.DB_NAME);
-   
-    // TODO: fix GAMES_COLLETION_NAME
-    const jobsCollection: mongoDB.Collection = db.collection(process.env.GAMES_COLLECTION_NAME);
-    
-    collections.jobs = jobsCollection;
-    
-    console.log(`Successfully connected to database: ${db.databaseName} and collection: ${jobsCollection.collectionName}`);
- }
+    try {
+        const query = { _id: new ObjectId(id) };
+        const job = (await collections.jobs?.findOne(query)) as unknown as Job;
+
+        if (job) {
+            res.status(200).send(job);
+        }
+    } catch (error) {
+        res.status(404).send(`Unable to find matching document with id: ${id}`);
+    }
+};
+
+const createJob = async (req: Request, res: Response) => {
+    try {
+        const newJob = req.body as Job;
+        const result = await collections.jobs?.insertOne(newJob);
+
+        result
+            ? res.status(201).send(`Successfully created a new job with id ${result.insertedId}`)
+            : res.status(500).send("Failed to create a new job.");
+    } catch (error) {
+        if (error instanceof Error) {
+            console.log(error);
+            res.status(400).send(error.message);
+        } else {
+            console.log('Unexpected error', error);
+            res.status(400).send(error);
+        }
+    }
+};
+
+export { getJob, createJob };
