+++ server/src/database/conn.ts
@@ -4,16 +4,15 @@ import * as dotenv from "dotenv";
 
 //TODO: Add your collections into here as you define them, then use them in your models
 // See https://www.mongodb.com/compatibility/using-typescript-with-mongodb-tutorial for specifics
-export const collections = {}
+export const collections: { jobs?: mongoDB.Collection } = {};
 
 const DB_CONN_STRING = process.env.DB_CONN_STRING || "mongodb://mongo:27017";
 const DB_NAME = process.env.DB_NAME || "default_db_name";
-
-
+const JOBS_COLLECTION_NAME = process.env.JOBS_COLLECTION_NAME || "default_jobs_collection_name";
 
 
 export async function connectToDatabase() {
-
+  dotenv.config();
 
   const client = new mongoDB.MongoClient(DB_CONN_STRING)
 
@@ -26,8 +25,11 @@ export async function connectToDatabase() {
     `Successfully connected to database: ${db.databaseName}`,
   );
 
-
-
+  const jobsCollection: mongoDB.Collection = db.collection(JOBS_COLLECTION_NAME);
+  collections.jobs = jobsCollection;
+  console.log(
+    `Successfully connected to collection: ${jobsCollection.collectionName}`
+  );
 
 }
